Introduce the concept of SyncEngine.
SyncEngine handles all functionality related to syncing to Dropbox. With
this chanage, Wiki no longer needs to deal directly with
Dropbox. Instead, it deals only with the filesystem.

Similarly, the filesystem does not need to know about
SyncEngine. SyncEngine listens to the events emitted by the filesystem
and manages syncing logic independently.

I made this refactoring primarily to move logic out of Wiki.